requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-util==1.0.17
drf-yasg==1.21.5
executing==1.2.0
fixtures-mongoengine==1.3.1
gevent==22.10.2
greenlet==2.0.2
huey==2.4.5
idna==3.4
importlib-resources==5.10.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mongoengine==0.26.0
packaging==23.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
PyJWT==2.6.0
pymongo==4.3.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
varname==0.10.0
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2
